discrete distribs doctests + delete Halton Owen
