+++ qmcpy/discrete_distribution/__init__.py
@@ -3,4 +3,3 @@ from .iid_std_gaussian import IIDStdGaussian
 from .lattice import Lattice
 from .digital_net_b2 import DigitalNetB2, Sobol, SobolSciPy
 from .halton import Halton
-from .korobov import Korobov
