--- qmcpy/discrete_distribution/halton/__init__.py
@@ -1,1 +1,0 @@
-from .halton import Halton
