+++ qmcpy/discrete_distribution/iid_std_gaussian.py
@@ -9,7 +9,15 @@ class IIDStdGaussian(DiscreteDistribution):
 
     >>> dd = IIDStdGaussian(dimension=2,seed=7)
     >>> dd.gen_samples(4)
+    array([[ 0.76412904,  1.38253623],
+           [ 1.59603668, -0.34658357],
+           [ 0.75751054, -1.75095847],
+           [ 0.14504699,  0.0736317 ]])
     >>> dd
+    IIDStdGaussian (DiscreteDistribution Object)
+        d               2^(1)
+        entropy         7
+        spawn_key       ()
     """
 
     def __init__(self, dimension=1, seed=None):
