+++ qmcpy/util/abstraction_functions.py
@@ -57,6 +57,6 @@ def _univ_repr(qmc_object, abc_class_name, attributes):
                     else: # float format
                         s += '%.3f' % val
             else:
-                s += '%s' % val
+                s += '%s' % str(val)
             string += '\n' + s.replace('\n', '\n    %-15s' % ' ')
     return string
